dbsite/requirements.txt
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.3
django-heroku==0.3.1
gunicorn==19.9.0
imgurpython==1.1.7
matplotlib==2.0.0
mysql-connector==2.1.6
numpy==1.12.1
psycopg2==2.7.6.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
whitenoise==4.1.2
